refactor: now uses Double instead of Float as per supervisor request

+++ src/main/java/de/featjar/feature/model/analysis/SimpleTreeProperties.java
@@ -14,17 +14,16 @@ public class SimpleTreeProperties {
 
     /**
      * @param tree: feature tree
-     * @return number of features directly below the root of this subtree.
+     * {@return number of features directly below the root of this subtree.}
      */
     public Result<Integer> topFeatures(IFeatureTree tree) {
-        // int childrenCount = tree.getRoot().getChildrenCount(); // if we should find a subtree's root automatically
-        int childrenCount = tree.getChildrenCount();
+        int childrenCount = tree.getChildrenCount(); // when doing computations: input the feature model
         return Result.of(childrenCount);
     }
 
     /**
      * @param tree: feature tree
-     * @return the number of features that have no child features
+     * {@return the number of features that have no child features}
      */
     public Result<Integer> leafFeaturesCounter(IFeatureTree tree) {
         return Trees.traverse(tree, new TreeLeafCounter());
@@ -32,7 +31,7 @@ public class SimpleTreeProperties {
 
     /**
      * @param tree: feature tree
-     * @return tree depth, meaning the longest path from this subtree's root to its most distant leaf node
+     * {@return tree depth, meaning the longest path from this subtree's root to its most distant leaf node}
      */
     public Result<Integer> treeDepth(IFeatureTree tree) {
         return Trees.traverse(tree, new TreeDepthCounter());
@@ -40,9 +39,9 @@ public class SimpleTreeProperties {
 
     /**
      * @param tree: feature tree
-     * @return average number of children that each node in the tree has, rounded to integer.
+     * {@return average number of children that each node in the tree has, rounded to integer.}
      */
-    public Result<Float> avgNumberOfChildren(IFeatureTree tree) {
+    public Result<Double> avgNumberOfChildren(IFeatureTree tree) {
         return Trees.traverse(tree, new TreeAvgChildrenCounter());
     }
